Kill git processes more gracefully on Linux/Mac
The kill.Kill() function uses SIGKILL to forcefully terminate processes. This is
not a good idea with git, since it might hit it at a time where it created an
index.lock file, which will then not be cleaned up. Improve this by sending the
command a SIGTERM signal instead.

I'm not changing the kill package here, because it is used by other projects
which might have different requirements than lazygit; so I'm fixing this locally
here. Also, I'm not changing Windows, it continues to use the kill.Kill code,
which forcefully terminates. This could still be a problem with regards to the
index.lock issue, but I'm just not sure if there's anything we can do, since
Windows doesn't support signals.