Prevent lazygit crashing for some xcbuild metadata
When nix brings in xcbuild as a transitive dependency on macOS,
if there are new metadata tags that the nix-specific version of
xcbuild doesn't support, then `git status` will write a warning
to stderr like: `warning: unhandled Platform key FamilyDisplayName`.
That stderr message causes lazygit to crash. My fix is to filter
out the warnings and then redirect them to the lazygit log

Fixes #5013