Remove the ExcludeFile action string
The action for ignoring a file says "Ignore or exclude file", but the action for
excluding a file says "Exclude file"; this asymmetry doesn't make sense. We
should either add a more specific "Ignore file" string for ignoring a file, or
use the more generic one for excluding too. I decided to do the latter, since
it's simpler, and I find it good enough.